pyproject.toml
[project]
name = "nf-processor-agent"
version = "0.1.0"
description = "Agente inteligente para processamento de notas fiscais"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Diogo Costa,Martina Froemming", email = ""}
]

dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "llama-index>=0.14.4",
    "llama-index-readers-file>=0.5.4",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "pandas>=2.2.3",
    "plotly>=6.3.1",
    "pydantic>=2.12.1",
    "pydot>=4.0.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "streamlit-extras>=0.7.8",
    "streamlit-option-menu>=0.4.0",
    "tqdm>=4.67.1",
    "xml2dict>=0.2.2",
    "xmltodict>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "isort>=5.12.0",
    "ipython>=8.17.2",
    "jupyter>=1.0.0",
]

test = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
]

docs = [
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/seu-usuario/nf-processor-agent"
Documentation = "https://nf-processor-agent.readthedocs.io"
Repository = "https://github.com/seu-usuario/nf-processor-agent"
Issues = "https://github.com/seu-usuario/nf-processor-agent/issues"

[tool.uv]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "W291",  # trailing whitespace
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
